[v0.10.8] [optimization] make new iterators
